+++ src/api/location.js
@@ -1,4 +1,5 @@
 const gql = require('../utils/api-client')
+const { handleError } = require('../utils/api-error')
 
 const LIST_QUERY = `
   query {
@@ -19,9 +20,7 @@ const list = async () => {
     const response = await gql.request(LIST_QUERY)
     return response.locations
   } catch (e) {
-    if (e.response && e.response.error) throw e.response
-    if (e.response && e.response.errors) throw e.response.errors
-    else throw e
+    return handleError(e)
   }
 }