+++ src/api/snapshot.js
@@ -1,9 +1,10 @@
 const gql = require('../utils/api-client')
+const { handleError } = require('../utils/api-error')
 
 const CREATE_MUTATION = `
   mutation CreateSnapshot(
     $site: String!
-    $ref: String
+    $ref: String!
   ) {
     createSnapshot(site: $site, ref: $ref) {
       iid
@@ -35,12 +36,13 @@ const LIST_QUERY = `
 
 const create = async ({ site, ref }) => {
   try {
-    const response = await gql.request(CREATE_MUTATION, { site, ref })
+    const response = await gql.request(CREATE_MUTATION, {
+      site,
+      ref: new String(ref)
+    })
     return response.createSnapshot
   } catch (e) {
-    if (e.response && e.response.error) throw e.response
-    if (e.response && e.response.errors) throw e.response.errors
-    else throw e
+    return handleError(e)
   }
 }
 
@@ -49,9 +51,7 @@ const list = async ({ site }) => {
     const response = await gql.request(LIST_QUERY, { site })
     return response.organisation.site.snapshots
   } catch (e) {
-    if (e.response && e.response.error) throw e.response
-    if (e.response && e.response.errors) throw e.response.errors
-    else throw e
+    return handleError(e)
   }
 }