+++ src/cli/site.js
@@ -1,6 +1,6 @@
 exports.command = 'site <command>'
 exports.desc = 'Tasks related to sites'
-exports.builder = function(yargs) {
+exports.builder = yargs => {
   return yargs.commandDir('site')
 }
-exports.handler = function(argv) {}
+exports.handler = argv => {}