+++ src/utils/api-error.js
@@ -1,0 +1,24 @@
+const handleError = err => {
+  if (err.response && err.response.error) throw err.response
+  if (err.response && err.response.errors) throw err.response.errors
+  else throw err
+}
+
+const humaniseError = apiError => {
+  if (
+    apiError[0] &&
+    apiError[0].problems &&
+    apiError[0].problems[0].explanation
+  ) {
+    return apiError[0].problems[0].explanation
+  } else if (apiError[0] && apiError[0].message) {
+    return apiError[0].message
+  } else {
+    return `An unknown error occurred ${JSON.stringify(apiError, null, 2)}`
+  }
+}
+
+module.exports = {
+  handleError,
+  humaniseError
+}